backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
supabase>=2.0.0
mem0ai>=0.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
cryptography>=43.0.0
astral>=3.2
cachetools>=5.5.0
numpy>=1.24.0
pandas>=2.0.0
fitparse>=1.2.0
lxml>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
